fix up types in BusView
also split up the multi-function files into multiple files
