+++ views/transportation/bus/get-set-of-stops-for-now.js
@@ -1,31 +1,31 @@
 // @flow
-import type {BusScheduleType} from './types'
-import moment from 'moment-timezone'
+import type {FancyBusTimeListType} from './types'
+import type moment from 'moment'
 import head from 'lodash/head'
 import last from 'lodash/last'
 import find from 'lodash/find'
 
 export default function getSetOfStopsForNow(
-  schedule: BusScheduleType,
-  now: typeof moment
-): (typeof moment)[] {
-  let times: string[]|void = find(schedule.times, times => {
-    const startTime = head(times)
-    const endTime = last(times)
+  scheduledMoments: FancyBusTimeListType[],
+  now: moment
+): FancyBusTimeListType {
+  let moments: ?Array<moment|false> = find(scheduledMoments, moments => {
+    const startTime = head(moments)
+    const endTime = last(moments)
     // Momentjs inclusivity: A [ indicates inclusion of a value. A ( indicates
     // exclusion. If the inclusivity parameter is used, both indicators must
     // be passed.
     return now.isBetween(startTime, endTime, null, '[]')
   })
 
-  if (times) {
-    return times
+  if (moments) {
+    return moments
   }
 
-  let firstBus = head(head(schedule.times))
+  let firstBus = head(head(scheduledMoments))
   if (now.isSameOrBefore(firstBus, 'minute')) {
-    return head(schedule.times)
+    return head(scheduledMoments)
   }
 
-  return last(schedule.times)
+  return last(scheduledMoments)
 }
