+++ views/transportation/bus/index.js
@@ -1,13 +1,21 @@
 // @flow
 import React from 'react'
 import {ScrollView} from 'react-native'
-import busInfo from '../../../data/bus-times.json'
+import defaultBusLines from '../../../data/bus-times.json'
 import type {BusLineType} from './types'
 import BusLineView from './bus-line'
 import moment from 'moment-timezone'
 const TIMEZONE = 'America/Winnipeg'
 
 export default class BusView extends React.Component {
+  static propTypes = {
+    busLines: React.PropTypes.arrayOf(React.PropTypes.object),
+  }
+
+  static defaultProps = {
+    busLines: defaultBusLines,
+  }
+
   state = {
     intervalId: 0,
     now: Date.now(),
@@ -16,11 +24,15 @@ export default class BusView extends React.Component {
   componentWillMount() {
     // This updates the screen every second, so that the "next bus" times are
     // updated without needing to leave and come back.
-    this.setState({intervalId: setInterval(this.updateTime, 1000)})
+    this.setState({intervalId: setInterval(this.updateTime, 5000)})
   }
 
   componentWillUnmount() {
     clearTimeout(this.state.intervalId)
+  }
+
+  props: {
+    busLines: BusLineType[]
   }
 
   updateTime = () => {
@@ -31,13 +43,14 @@ export default class BusView extends React.Component {
     // const now = moment.tz('6:53am', 'h:mma', true, TIMEZONE)
     // now.dayOfYear(moment.tz(TIMEZONE).dayOfYear())
     const now = moment.tz(TIMEZONE)
+    const busLines = this.props.busLines
 
     return (
       <ScrollView>
-        {busInfo.map((busLine: BusLineType, i) =>
+        {busLines.map((busLine: BusLineType, i) =>
           <BusLineView
             key={busLine.line}
-            style={{marginBottom: i < busInfo.length - 1 ? 5 : 15}}
+            style={{marginBottom: i < busLines.length - 1 ? 5 : 15}}
             line={busLine}
             now={now}
           />)}
